requirements.txt
altgraph==0.17.4
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
docutils==0.22.2
filetype==1.2.0
fonttools==4.60.1
idna==3.11
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.9
macholib==1.16.3
matplotlib==3.10.7
numpy==2.2.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
pyserial>=3.5